+++ src/br/com/g5/agendacontatos/contato/Contato.java
@@ -1,6 +1,6 @@
 package br.com.g5.agendacontatos.contato;
 
-import br.com.g5.agendacontatos.Filme;
+import br.com.g5.agendacontatos.filme.Filme;
 import br.com.g5.agendacontatos.telefone.Telefone;
 import br.com.g5.agendacontatos.util.Predicados;
 import br.com.g5.agendacontatos.util.Verificador;
