+++ src/br/com/g5/agendacontatos/filme/Filme.java
@@ -1,4 +1,4 @@
-package br.com.g5.agendacontatos;
+package br.com.g5.agendacontatos.filme;
 
 public class Filme {
     private String nome;
